+++ src/printer/preprocess/augment-with-css-properties.ts
@@ -95,9 +95,7 @@ function getCssDisplay(
         case 'ignore':
           return 'block';
         default: {
-          return (
-            CSS_DISPLAY_LAVA_TAGS[node.name] || CSS_DISPLAY_LAVA_DEFAULT
-          );
+          return CSS_DISPLAY_LAVA_TAGS[node.name] || CSS_DISPLAY_LAVA_DEFAULT;
         }
       }