+++ src/printer/print/lava.ts
@@ -218,7 +218,7 @@ function printNamedLavaBlockStart(
     }
 
     case NamedTags.if:
-    case NamedTags.elsif:
+    case NamedTags.elseif:
     case NamedTags.unless: {
       const trailingWhitespace = [
         NodeTypes.Comparison,