+++ src/printer/utils/node.ts
@@ -81,9 +81,7 @@ export function isTextLikeNode(
   return !!node && node.type === NodeTypes.TextNode;
 }
 
-export function isLavaNode(
-  node: LavaHtmlNode | undefined,
-): node is LavaNode {
+export function isLavaNode(node: LavaHtmlNode | undefined): node is LavaNode {
   return !!node && LavaNodeTypes.includes(node.type as any);
 }