+++ src/types.ts
@@ -76,7 +76,7 @@ export enum NamedTags {
   cycle = 'cycle',
   decrement = 'decrement',
   echo = 'echo',
-  elsif = 'elsif',
+  elseif = 'elseif',
   for = 'for',
   form = 'form',
   if = 'if',
@@ -254,10 +254,7 @@ export type LavaRawTag = Augmented<AST.LavaRawTag, AllAugmentations>;
 export type LavaTag = Augmented<AST.LavaTag, AllAugmentations>;
 export type LavaTagNamed = Augmented<AST.LavaTagNamed, AllAugmentations>;
 export type LavaBranch = Augmented<AST.LavaBranch, AllAugmentations>;
-export type LavaBranchNamed = Augmented<
-  AST.LavaBranchNamed,
-  AllAugmentations
->;
+export type LavaBranchNamed = Augmented<AST.LavaBranchNamed, AllAugmentations>;
 export type LavaDrop = Augmented<AST.LavaDrop, AllAugmentations>;
 export type HtmlNode = Augmented<AST.HtmlNode, AllAugmentations>;
 export type HtmlTag = Exclude<HtmlNode, HtmlComment>;
@@ -289,8 +286,5 @@ export type AttrDoubleQuoted = Augmented<
 >;
 export type AttrUnquoted = Augmented<AST.AttrUnquoted, AllAugmentations>;
 export type AttrEmpty = Augmented<AST.AttrEmpty, AllAugmentations>;
-export type LavaExpression = Augmented<
-  AST.LavaExpression,
-  AllAugmentations
->;
+export type LavaExpression = Augmented<AST.LavaExpression, AllAugmentations>;
 export type TextNode = Augmented<AST.TextNode, AllAugmentations>;