+++ test/lava-style-tag/index.spec.ts
@@ -1,6 +1,0 @@
-import { assertFormattedEqualsFixed } from '../test-helpers';
-import * as path from 'path';
-
-describe(`Unit: ${path.basename(__dirname)}`, () => {
-  assertFormattedEqualsFixed(__dirname);
-});